+++ modules/photo/src/tonemap.cpp
@@ -505,7 +505,7 @@ protected:
     void calculateSum(std::vector<Mat>& x_contrast, std::vector<Mat>& y_contrast, Mat& sum)
     {
         sum = Mat::zeros(x_contrast[x_contrast.size() - 1].size(), CV_32F);
-        for(int i = x_contrast.size() - 1; i >= 0; i--)
+        for(int i = (int)x_contrast.size() - 1; i >= 0; i--)
         {
             Mat grad_x, grad_y;
             getGradient(x_contrast[i], grad_x, 1);
